refactored tokens to split up by modules
